refactor(dashboard): G-series — extract memory_search + pilot_proxy (SR-57..58)
codec_dashboard.py: 2,187 → 2,018 LOC (-169 / -7.7%).
Cumulative (since main pre-B6): 3,912 → 2,018 (-48.4%).

New modules:
  routes/memory_search.py    /api/memory/search (143 LOC, 4 sources:
                             voice FTS + qchat + vibe + flash)
  routes/pilot_proxy.py      /api/pilot/{path:path} (HTTPS→localhost:8094)

Test surface (+7 new pins):
  - 2 endpoint-registered pins
  - routes-module-exports-router pin
  - memory_search must still hit all 4 backends (codec_memory + routes.qchat +
    routes.vibe + sessions table) — guards against quiet source removal
  - pilot_proxy must still target localhost:8094 + accept all 4 verbs
  - dashboard MUST NOT redefine either decorator
  - LOC floor < 2,100

Full suite: 2,046 passed / 77 skipped (was 2,039 in F-series). +7 net.

Drive-by ruff F401: removed unused `Response` import in dashboard now that the
pilot proxy is gone (it was the only consumer).

Remaining route-extraction opportunities deferred:
  - /api/command (180 LOC, safety-critical — is_dangerous + skill hijack + consent)
  - /api/chat chat_completion (608 LOC streaming + post-LLM [SKILL:...] tag)
  - /api/services/status (small, reads dashboard-owned _bg_status / _bg_tasks globals)
  - page renderers + startup/shutdown hooks + background daemons
